# backend/docs/openapi.yaml
openapi: '3.0.0'
info:
  title: Contract 961.kz API
  description: |
    API документация для системы электронного подписания документов Contract 961.kz
    
    ## Аутентификация
    Большинство эндпоинтов требуют JWT токен в заголовке Authorization:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Общий лимит: 100 запросов/минуту
    - OTP отправка: 3 запроса/15 минут
    - OTP верификация: 10 запросов/5 минут
  version: '1.0.0'
  contact:
    name: Contract 961.kz
    email: support@961.kz
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.961.kz
    description: Production
...
paths:
...
  /api/requests:
    get:
      tags: [Requests]
      summary: Получить список заявок
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, viewed, code_sent, signed, rejected]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список заявок
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/Request'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer

    post:
      tags: [Requests]
      summary: Создать новую заявку
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [clientName, clientPhone, documentName, file]
              properties:
                clientName:
                  type: string
                clientPhone:
                  type: string
                clientEmail:
                  type: string
                  format: email
                documentName:
                  type: string
                notes:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
  /api/requests/{id}:
    get:
      tags: [Requests]
      summary: Получить заявку по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Детали заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDetail'

    delete:
      tags: [Requests]
      summary: Удалить заявку
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Заявка удалена
...
  /api/requests/export:
    get:
      tags: [Requests]
      summary: Экспорт заявок в Excel
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Excel файл
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                properties:
                  empty:
                    type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Request:
      type: object
      properties:
        id:
          type: string
        displayId:
          type: string
        documentName:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        status:
          type: string
          enum: [draft, sent, viewed, code_sent, signed, rejected]
        createdAt:
          type: string
          format: date-time
        signedAt:
          type: string
          format: date-time
          nullable: true
    RequestDetail:
      allOf:
        - $ref: '#/components/schemas/Request'
        - type: object
          properties:
            signUrl:
              type: string
            document:
              type: object
              properties:
                id:
                  type: string
                originalHash:
                  type: string